services/process_worker/requirements.txt
redis==5.0.1
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
fastapi==0.104.1
uvicorn==0.24.0.post1
pydantic==2.5.2
